html/semantics/forms/the-input-element/password.html

<!DOCTYPE html>
<html>
  <head>
    <title>Password input element</title>
    <link rel="author" title="Shaoning Zeng" href="mailto:shaoning.zeng@gmail.com">
    <link rel="help" href="http://www.w3.org/TR/html5/forms.html#password-state-(type=password)">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <h1>Password input element</h1>
    <div style="display: none">

    <input id="password" type="password" />

    <input id="password_with_value" type="password" value="foo123！@#￥%……*（）"/>
    <input id="password_with_LF" type="password" value="0&#x000A;1" />
    <input id="password_with_CR" type="password" value="0&#x000D;1" />

    </div>
    <div id="log"></div>
  <script type="text/javascript">

    var element = document.getElementById("password");
    test(
      function() {
        assert_equals(element.type, "password");
        assert_equals(element.value, "");
        assert_equals(document.getElementById("password_with_value").type, "password");
        assert_equals(document.getElementById("password_with_value").value, "foo123！@#￥%……*（）");
      }, "Value returns the current value for password input.");

    test(
      function() {
        assert_equals(document.getElementById("password_with_LF").value, "01");
        assert_equals(document.getElementById("password_with_CR").value, "01");
      }, 'The value attribute, if specified, must have a value that contains no "LF" (U+000A) or "CR" (U+000D) characters.'); 

    test(
      function() {
        element.value = "0\u000A1";
        assert_equals(element.value, "01");
        element.value = "0\u000A\u000A1\u000A\u000A";
        assert_equals(element.value, "01");
        element.value = "0\u000D1";
        assert_equals(element.value, "01");
        element.value = "0\u000D\u000D1\u000D\u000D";
        assert_equals(element.value, "01");
      }, 'The value sanitization algorithm is as follows: Strip line breaks from the value.'); 

  </script>
  </body>
</html>
